make a proper infra package
